+++ tests/settings.py
@@ -98,9 +98,9 @@ FLOW_EXECUTOR = {
     'DATA_DIR': os.path.join(PROJECT_ROOT, '.test_data'),
     'UPLOAD_DIR': os.path.join(PROJECT_ROOT, '.test_upload'),
 }
-# Set custom executor command if set via environment variable
-if 'RESOLWE_EXECUTOR_COMMAND' in os.environ:
-    FLOW_EXECUTOR['COMMAND'] = os.environ['RESOLWE_EXECUTOR_COMMAND']
+# Set custom Docker command if set via environment variable
+if 'RESOLWE_DOCKER_COMMAND' in os.environ:
+    FLOW_DOCKER_COMMAND = os.environ['RESOLWE_DOCKER_COMMAND']
 FLOW_API = {
     'PERMISSIONS': 'resolwe.permissions.permissions',
 }